[ENH] Move remaining watershed code into pipeline component
